requirements.txt
fastapi
uvicorn
python-multipart
pillow
opencv-python
numpy
pandas
scikit-learn
torch
torchvision